fix: consolidate and remove duplicate consts
