+++ apps/content-publishing-api/src/api.service.ts
@@ -16,7 +16,7 @@ import {
   AttachmentType,
 } from '#types/dtos/content-publishing';
 import { IRequestJob, IAssetMetadata, IAssetJob } from '#types/interfaces/content-publishing';
-import { REQUEST_QUEUE_NAME, ASSET_QUEUE_NAME } from '#types/constants/content-publishing.queue.constants';
+import { REQUEST_QUEUE_NAME, ASSET_QUEUE_NAME } from '#types/constants';
 import { calculateIpfsCID } from '#content-publishing-lib/utils/ipfs';
 import {
   getAssetMetadataKey,
