+++ apps/content-publishing-api/src/controllers/v1/development.controller.v1.ts
@@ -22,7 +22,7 @@ import {
   ReactionAnnouncement,
   TombstoneAnnouncement,
 } from '#types/interfaces/content-publishing';
-import * as QueueConstants from '#types/constants/content-publishing.queue.constants';
+import * as QueueConstants from '#types/constants/queue.constants';
 import { calculateDsnpHash } from '#content-publishing-lib/utils/ipfs';
 import { IpfsService } from '#content-publishing-lib/utils/ipfs.client';
 import { AnnouncementTypeDto } from '#types/dtos/content-publishing';
