+++ apps/content-publishing-worker/src/batch_announcer/batch.announcer.service.ts
@@ -4,9 +4,9 @@ import { Job, Queue } from 'bullmq';
 import { SchedulerRegistry } from '@nestjs/schedule';
 import { BatchAnnouncer } from './batch.announcer';
 import { BaseConsumer } from '../BaseConsumer';
-import { BATCH_QUEUE_NAME, PUBLISH_QUEUE_NAME } from '#types/constants/content-publishing.queue.constants';
+import { BATCH_QUEUE_NAME, PUBLISH_QUEUE_NAME } from '#types/constants/queue.constants';
 import { IBatchAnnouncerJobData } from '#types/interfaces/content-publishing';
-import { CAPACITY_EPOCH_TIMEOUT_NAME } from '#types/constants/content-publishing-constants';
+import { CAPACITY_EPOCH_TIMEOUT_NAME } from '#types/constants';
 
 @Injectable()
 @Processor(BATCH_QUEUE_NAME, {
