+++ apps/content-publishing-worker/src/batching_processor/workers/reaction.worker.ts
@@ -1,5 +1,5 @@
 import { Announcement } from '#types/interfaces/content-publishing';
-import { REACTION_QUEUE_NAME } from '#types/constants/content-publishing.queue.constants';
+import { REACTION_QUEUE_NAME } from '#types/constants/queue.constants';
 import { Processor, OnWorkerEvent } from '@nestjs/bullmq';
 import { Injectable, OnApplicationBootstrap } from '@nestjs/common';
 import { Job } from 'bullmq';
