+++ apps/content-publishing-worker/src/request_processor/request.processor.service.ts
@@ -4,7 +4,7 @@ import { DelayedError, Job } from 'bullmq';
 import { MILLISECONDS_PER_SECOND } from 'time-constants';
 import { ConfigService } from '#content-publishing-lib/config';
 import { IRequestJob } from '#types/interfaces/content-publishing';
-import { REQUEST_QUEUE_NAME } from '#types/constants/content-publishing.queue.constants';
+import { REQUEST_QUEUE_NAME } from '#types/constants/queue.constants';
 import { IpfsService } from '#content-publishing-lib/utils/ipfs.client';
 import { BaseConsumer } from '../BaseConsumer';
 import { DsnpAnnouncementProcessor } from './dsnp.announcement.processor';
