+++ libs/content-watcher-lib/src/pubsub/announcers/profile.ts
@@ -1,7 +1,7 @@
 import { Processor } from '@nestjs/bullmq';
 import { Injectable } from '@nestjs/common';
 import { Job } from 'bullmq';
-import * as QueueConstants from '#types/constants/content-watcher.queue.constants';
+import * as QueueConstants from '#types/constants/queue.constants';
 import { BaseConsumer } from '../../utils/base-consumer';
 import { PubSubService } from '../pubsub.service';
 import { AnnouncementResponse } from '#types/content-announcement';
