+++ libs/content-watcher-lib/src/scanner/scanner.service.ts
@@ -10,12 +10,8 @@ import { MILLISECONDS_PER_SECOND } from 'time-constants';
 import { Queue } from 'bullmq';
 import { AppConfigService } from '../config/config.service';
 import { BlockchainService } from '../blockchain/blockchain.service';
-import * as QueueConstants from '#types/constants/content-watcher.queue.constants';
-import {
-  EVENTS_TO_WATCH_KEY,
-  LAST_SEEN_BLOCK_NUMBER_SCANNER_KEY,
-  REGISTERED_WEBHOOK_KEY,
-} from '#types/constants/content-watcher.constants';
+import * as QueueConstants from '#types/constants/queue.constants';
+import { EVENTS_TO_WATCH_KEY, LAST_SEEN_BLOCK_NUMBER_SCANNER_KEY, REGISTERED_WEBHOOK_KEY } from '#types/constants';
 import { ChainWatchOptionsDto } from '#types/dtos/content-watcher/chain.watch.dto';
 import * as RedisUtils from '#content-watcher-lib/utils/redis';
 import { ChainEventProcessorService } from '../blockchain/chain-event-processor.service';
