+++ libs/types/src/constants/blockchain-constants.ts
@@ -1,3 +1,4 @@
-export namespace BlockchainConstants {
-  export const SECONDS_PER_BLOCK = 12;
-}
+/**
+ * Number of seconds to create a block in Frequency chain
+ */
+export const SECONDS_PER_BLOCK = 12;
