--- libs/types/src/constants/content-publishing-constants.ts
@@ -1,9 +1,0 @@
-/**
- * Name of timeout event used for in memory scheduler
- */
-export const CAPACITY_EPOCH_TIMEOUT_NAME = 'capacity-epoch-timeout';
-
-/**
- * Hash set key containing ITxStatus values for submitted chain transactions we are watching for completion
- */
-export const TXN_WATCH_LIST_KEY = 'txnWatchList';
