+++ libs/types/src/enums/environment-type.enum.ts
@@ -1,0 +1,6 @@
+// eslint-disable-next-line no-shadow
+export enum EnvironmentType {
+  MAIN_NET = 'mainnet',
+  ROCOCO = 'rococo',
+  DEV = 'dev',
+}
